Refactor n2c_server.rs (3,303 lines) into per-protocol modules
Split the monolithic N2C server file into an n2c/ directory with
per-protocol modules for better maintainability:

- n2c/mod.rs: N2CServer struct, handshake, connection dispatch, shared types
- n2c/chain_sync.rs: LocalChainSync protocol handler
- n2c/state_query.rs: LocalStateQuery protocol handler + CBOR encoding
- n2c/tx_submission.rs: LocalTxSubmission protocol handler
- n2c/tx_monitor.rs: LocalTxMonitor protocol handler

Pure refactoring with no behavioral changes. All 148 tests pass.
