+++ crates/torsten-network/src/lib.rs
@@ -1,8 +1,8 @@
 pub mod client;
 pub mod miniprotocols;
 pub mod multiplexer;
+pub mod n2c;
 pub mod n2c_client;
-pub mod n2c_server;
 pub mod n2n_server;
 pub mod peer;
 pub mod peer_manager;
@@ -15,8 +15,8 @@ pub use miniprotocols::peersharing::{
     request_peers_from, PeerAddress, PeerSharingMessage, PeerSharingState,
 };
 pub use miniprotocols::txsubmission::{TxSubmissionClient, TxSubmissionError, TxSubmissionStats};
+pub use n2c::{N2CServer, TxValidator};
 pub use n2c_client::N2CClient;
-pub use n2c_server::{N2CServer, TxValidator};
 pub use n2n_server::{BlockAnnouncement, BlockProvider, N2NServer, RollbackAnnouncement, TipInfo};
 pub use peer::PeerConnection;
 pub use peer_manager::{DiffusionMode, PeerManager, PeerManagerConfig, PeerPerformance};
